refactor: update address row component with labeled into a component, reuse in send and cfp screen
